pyproject.toml
[project]
name = "cardagentpoc"
version = "0.1.0"
description = "CardHero Last-Mile Engine — AI-assisted Pokemon card deal hunter"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "praw>=7.7.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
